# backend/src/main/java/com/roman/task_tracker/role/RoleResponse.java
package com.roman.task_tracker.role;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.UUID;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class RoleResponse {
    private String name;
    private UUID id;


}